requirements.txt
fastapi[standard]
python-jose
passlib
bycrypt
pymongo